+++ src/modules/Domloading.js
@@ -1,0 +1,4 @@
+
+document.addEventListener('DOMContentLoaded', () => {
+  
+});
